reviews of the sale representative with there clients

+++ src/main/java/com/spring/jwt/dto/FeedbackRequestDto.java
@@ -19,5 +19,5 @@ public class FeedbackRequestDto {
     @Max(5)
     private Integer rating;
 
-    private Long businessId;
+    private Integer businessId;
 }
